+++ src/index.js
@@ -2,9 +2,14 @@ import React, { lazy, Suspense } from 'react';
 import PropTypes from 'prop-types';
 import { Switch } from 'react-router-dom';
 import { Route } from '@folio/stripes/core';
-import { CommandList, HasCommand, Layout } from '@folio/stripes/components';
+import {
+  CommandList,
+  HasCommand,
+  Layout,
+  checkScope,
+  defaultKeyboardShortcuts,
+} from '@folio/stripes/components';
 
-import { keyboardCommands } from '@folio/stripes-erm-components';
 import css from './index.css';
 
 const AgreementsRoute = lazy(() => import('./routes/AgreementsRoute'));
@@ -81,10 +86,6 @@ class App extends React.Component {
     },
   ];
 
-  checkScope = () => {
-    return document.body.contains(document.activeElement);
-  }
-
   render() {
     const { actAs, match: { path } } = this.props;
 
@@ -97,10 +98,10 @@ class App extends React.Component {
     }
 
     return (
-      <CommandList commands={keyboardCommands}>
+      <CommandList commands={defaultKeyboardShortcuts}>
         <HasCommand
           commands={this.shortcuts}
-          isWithinScope={this.checkScope}
+          isWithinScope={checkScope}
           scope={document.body}
         >
           <div className={css.container}>